+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+0.6.0 (unreleased)
+++++++++++++++++++
+
+* LatestPostsPlugin tags field has been changed to a plain TaggableManager field.
+  A migration is in place to move the data, but backup your data first.
+
 0.5.0 (2015-08-09)
 ++++++++++++++++++